[Feat] 리뷰 CURD API 구현 (#3)
* [Feat] 리뷰 작성 API 구현 (#3)

* [Feat] 리뷰 조회 API 구현 (#3)

* [Fix] 식당 조회 API 응답 형식 변경 (#1)

* [Refactor] 리뷰 조회 API 코드 리팩토링 (#3)

* [Fix] query 타입 변경 (#3)

+++ src/main/java/com/dsfs/dsfs/domain/Review.java
@@ -14,7 +14,7 @@ public class Review extends BaseEntity {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long reviewId;
-    private Double score;
+    private Float score;
     private String content;
 
     @ManyToOne(fetch = FetchType.LAZY)
